Task #62: refactored an api
