+++ src/LingoMq.sln
@@ -27,8 +27,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Achievements", "Achievement
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Topics", "Topics", "{8F1AFB24-2636-45D5-AED2-B40E7C6CC02E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Topics.Api", "Services\Topics\Topics.Api\Topics.Api.csproj", "{78A62A89-ABFE-4ED5-ABCE-A0C2451FBCB8}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AppStatistics", "AppStatistics", "{66D24938-EF5A-4F0F-A041-492653D06668}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppStatistics.Api", "Services\AppStatistics\AppStatistics.Api\AppStatistics.Api.csproj", "{C8425872-AE91-4B9E-80BE-55DA6EFD372F}"
@@ -109,7 +107,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DataAccess", "DataAccess", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DataAccess.Dapper", "Services\Topics\DataAccess\DataAccess.Dapper\DataAccess.Dapper.csproj", "{F257F69D-ADA4-41FC-B9E4-8EF7757938B8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Application", "Services\Topics\Application\Application.csproj", "{05434716-8AD7-41D8-A740-E5EE495BF081}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "Services\Topics\Application\Application.csproj", "{05434716-8AD7-41D8-A740-E5EE495BF081}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApi", "Services\Topics\WebApi\WebApi.csproj", "{7B8BE181-B962-4A16-83D1-CE9AB9F070BF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -129,10 +129,6 @@ Global
 		{9EFF7B1E-EAE6-4388-A727-616C36809DDD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9EFF7B1E-EAE6-4388-A727-616C36809DDD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9EFF7B1E-EAE6-4388-A727-616C36809DDD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{78A62A89-ABFE-4ED5-ABCE-A0C2451FBCB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{78A62A89-ABFE-4ED5-ABCE-A0C2451FBCB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{78A62A89-ABFE-4ED5-ABCE-A0C2451FBCB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{78A62A89-ABFE-4ED5-ABCE-A0C2451FBCB8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C8425872-AE91-4B9E-80BE-55DA6EFD372F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C8425872-AE91-4B9E-80BE-55DA6EFD372F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C8425872-AE91-4B9E-80BE-55DA6EFD372F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -261,6 +257,10 @@ Global
 		{05434716-8AD7-41D8-A740-E5EE495BF081}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{05434716-8AD7-41D8-A740-E5EE495BF081}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{05434716-8AD7-41D8-A740-E5EE495BF081}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7B8BE181-B962-4A16-83D1-CE9AB9F070BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7B8BE181-B962-4A16-83D1-CE9AB9F070BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7B8BE181-B962-4A16-83D1-CE9AB9F070BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7B8BE181-B962-4A16-83D1-CE9AB9F070BF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -276,7 +276,6 @@ Global
 		{AC5C0041-5671-4533-B6A7-74BB1127A3F2} = {EFBD5049-E4DF-420E-A693-5FDBA3D4E215}
 		{F488E10A-94B3-4BA2-96A0-0BE98FDAC4A0} = {EFBD5049-E4DF-420E-A693-5FDBA3D4E215}
 		{8F1AFB24-2636-45D5-AED2-B40E7C6CC02E} = {EFBD5049-E4DF-420E-A693-5FDBA3D4E215}
-		{78A62A89-ABFE-4ED5-ABCE-A0C2451FBCB8} = {8F1AFB24-2636-45D5-AED2-B40E7C6CC02E}
 		{66D24938-EF5A-4F0F-A041-492653D06668} = {EFBD5049-E4DF-420E-A693-5FDBA3D4E215}
 		{C8425872-AE91-4B9E-80BE-55DA6EFD372F} = {66D24938-EF5A-4F0F-A041-492653D06668}
 		{73F223E8-5A4E-48BA-9F71-70C3E2E62B35} = {66D24938-EF5A-4F0F-A041-492653D06668}
@@ -317,6 +316,7 @@ Global
 		{906FFA71-7A1C-4943-BCEB-543BDE3BE17C} = {8F1AFB24-2636-45D5-AED2-B40E7C6CC02E}
 		{F257F69D-ADA4-41FC-B9E4-8EF7757938B8} = {906FFA71-7A1C-4943-BCEB-543BDE3BE17C}
 		{05434716-8AD7-41D8-A740-E5EE495BF081} = {8F1AFB24-2636-45D5-AED2-B40E7C6CC02E}
+		{7B8BE181-B962-4A16-83D1-CE9AB9F070BF} = {8F1AFB24-2636-45D5-AED2-B40E7C6CC02E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2746943F-0DF5-4BB2-A360-520130DFE182}
