--- src/Services/Topics/Topics.Api/NLog.config
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-    <targets>
-        <target name="console" xsi:type="Console" 
-         layout="[${uppercase:${level}}] ${message} \t ${longdate}" />
-        <target name="file" xsi:type="File" fileName="${basedir}/logs/${shortdate}.txt" 
-            layout="[${uppercase:${level}}] ${message} \t ${longdate}" />
-    </targets>
-    <rules>
-        <logger name="*" levels="Info,Error" writeTo="console" />
-        <logger name="*" levels="Info,Debug,Error" writeTo="file" />
-    </rules>
-</nlog>
