--- src/Services/Topics/Topics.Api/Program.cs
@@ -1,169 +1,0 @@
-using FluentMigrator.Runner;
-using MassTransit;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.OpenApi.Models;
-using NLog.Extensions.Logging;
-using Npgsql;
-using System.Data;
-using System.Reflection;
-using System.Text;
-using Topics.Api.Middlewares;
-using Topics.Api.Services;
-using Topics.BusinessLayer.Contracts;
-using Topics.BusinessLayer.MassTransit.Consumers.IdentityConsumers;
-using Topics.BusinessLayer.Services;
-using Topics.BusinessLayer.Services.Repositories;
-
-var builder = WebApplication.CreateBuilder(args);
-builder.Configuration
-    .SetBasePath(Directory.GetCurrentDirectory())
-    .AddJsonFile("appsettings.json", true, true)
-    .AddEnvironmentVariables();
-
-// Add Logging (NLog)
-builder.Services.AddLogging(loggingBuilder =>
-{
-    loggingBuilder.AddNLog();
-});
-
-// DataLayer
-builder.Services.AddTransient<IDbConnection>((sp) => new NpgsqlConnection(builder.Configuration["ConnectionStrings:Dev:Topics"]));
-builder.Services.AddTransient<ILanguageRepository, LanguageRepository>();
-builder.Services.AddTransient<ITopicLevelRepository, TopicLevelRepository>();
-builder.Services.AddTransient<ITopicRepository, TopicRepository>();
-builder.Services.AddTransient<ITopicStatisticsRepository, TopicStatisticsRepository>();
-builder.Services.AddTransient<ITopicStatisticsTypeRepository, TopicStatisticsTypeRepository>();
-builder.Services.AddTransient<IUserRepository, UserRepository>();
-builder.Services.AddTransient<IUnitOfWork, UnitOfWork>();
-builder.Services.AddTransient<IDatabaseDataMigrator, DatabaseDataMigrator>();
-
-// Authentication (current: JWT)
-builder.Services.AddAuthentication(x =>
-{
-    x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-})
-    .AddJwtBearer(o =>
-    {
-        o.TokenValidationParameters = new TokenValidationParameters
-        {
-            ValidateIssuer = true,
-            ValidIssuer = builder.Configuration["JWT:Issuer"],
-            ValidateAudience = true,
-            ValidAudience = builder.Configuration["JWT:Audience"],
-            ValidateLifetime = true,
-            ValidateIssuerSigningKey = true,
-            IssuerSigningKey = new SymmetricSecurityKey(
-                Encoding.UTF8.GetBytes(builder.Configuration["JWT:Key"]!))
-        };
-    });
-
-// Migrations
-builder.Services.AddFluentMigratorCore()
-        .ConfigureRunner(cr => cr
-        .AddPostgres()
-        .WithGlobalConnectionString(builder.Configuration["ConnectionStrings:Dev:Topics"])
-        .ScanIn(Assembly.GetExecutingAssembly()).For.Migrations());
-
-// MassTransit
-builder.Services.AddMassTransit(x =>
-{
-    x.SetKebabCaseEndpointNameFormatter();
-    x.AddDelayedMessageScheduler();
-    x.AddConsumer<TopicsDeleteUserConsumer>();
-    x.AddConsumer<TopicsUpdateUserConsumer>();
-    x.AddConsumer<TopicsCreateUserConsumer>();
-    x.UsingRabbitMq((context, cfg) =>
-    {
-        cfg.Host(builder.Configuration["RabbitMq:Uri"]!, "/", h =>
-        {
-            h.Username(builder.Configuration["RabbitMq:UserName"]);
-            h.Password(builder.Configuration["RabbitMq:Password"]);
-        });
-
-        cfg.ReceiveEndpoint(typeof(TopicsDeleteUserConsumer).Name.ToLower(), endpoint =>
-        {
-            endpoint.ConfigureConsumer<TopicsDeleteUserConsumer>(context);
-        });
-        cfg.ReceiveEndpoint(typeof(TopicsUpdateUserConsumer).Name.ToLower(), endpoint =>
-        {
-            endpoint.ConfigureConsumer<TopicsUpdateUserConsumer>(context);
-        });
-        cfg.ReceiveEndpoint(typeof(TopicsCreateUserConsumer).Name.ToLower(), endpoint =>
-        {
-            endpoint.ConfigureConsumer<TopicsCreateUserConsumer>(context);
-        });
-        cfg.ClearSerialization();
-        cfg.UseRawJsonSerializer();
-        cfg.ConfigureEndpoints(context);
-    });
-});
-
-// Swagger Auth interface
-builder.Services.AddSwaggerGen(c =>
-{
-    c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-    {
-        Description = @"JWT Authorization header using the Bearer scheme. \r\n\r\n 
-                      Enter 'Bearer' [space] and then your token in the text input below.
-                      \r\n\r\nExample: 'Bearer 12345abcdef'",
-        Name = "Authorization",
-        In = ParameterLocation.Header,
-        Type = SecuritySchemeType.ApiKey,
-        Scheme = "Bearer",
-        BearerFormat = "JWT",
-    });
-
-    c.AddSecurityRequirement(new OpenApiSecurityRequirement()
-      {
-        {
-          new OpenApiSecurityScheme
-          {
-            Reference = new OpenApiReference
-              {
-                Type = ReferenceType.SecurityScheme,
-                Id = "Bearer"
-              },
-              Scheme = "Jwt",
-              Name = "Bearer",
-              In = ParameterLocation.Header,
-
-            },
-             new string[] {}
-          }
-        });
-});
-
-builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.UseMiddleware<ExceptionHandlerMiddleware>();
-
-// Auto migrations
-using (var serviceScope = app.Services.CreateScope())
-{
-    var services = serviceScope.ServiceProvider;
-    var databaseMigrator = services.GetRequiredService<IDatabaseDataMigrator>();
-    var runner = services.GetRequiredService<IMigrationRunner>();
-
-    runner.MigrateUp();
-    await databaseMigrator.MigrateAsync();
-}
-
-app.Run();
