--- src/Services/Topics/Topics.Api/Services/IDatabaseDataMigrator.cs
@@ -1,10 +1,0 @@
-﻿namespace Topics.Api.Services
-{
-    public interface IDatabaseDataMigrator
-    {
-        Task AddLanguagesAsync();
-        Task AddTopicStatisticsTypeAsync();
-        Task AddTopicLevelAsync();
-        Task MigrateAsync();
-    }
-}
