+++ src/Services/Topics/WebApi/Extensions/SwaggerDependencyInjection.cs
@@ -1,0 +1,45 @@
+﻿using Microsoft.OpenApi.Models;
+
+namespace Microsoft.Extensions.DependencyInjection;
+
+public static class SwaggerDependencyInjection
+{
+    public static IServiceCollection AddSwagger(this IServiceCollection services)
+    {
+        services.AddSwaggerGen(c =>
+        {
+            c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
+            {
+                Description = @"JWT Authorization header using the Bearer scheme. \r\n\r\n 
+                      Enter 'Bearer' [space] and then your token in the text input below.
+                      \r\n\r\nExample: 'Bearer 12345abcdef'",
+                Name = "Authorization",
+                In = ParameterLocation.Header,
+                Type = SecuritySchemeType.ApiKey,
+                Scheme = "Bearer",
+                BearerFormat = "JWT",
+            });
+
+            c.AddSecurityRequirement(new OpenApiSecurityRequirement()
+      {
+        {
+          new OpenApiSecurityScheme
+          {
+            Reference = new OpenApiReference
+              {
+                Type = ReferenceType.SecurityScheme,
+                Id = "Bearer"
+              },
+              Scheme = "Jwt",
+              Name = "Bearer",
+              In = ParameterLocation.Header,
+
+            },
+             new string[] {}
+          }
+        });
+        });
+
+        return services;
+    }
+}
